+++ games/tanks/src/TensorFlow/Models/Layers/RoleEncodingLayer.ts
